엑추에이터 도메인 리펙토링
