+++ legend-momo-city/src/main/java/com/wanted/momocity/enrollment/application/service/EnrollmentCommandService.java
@@ -9,7 +9,6 @@ import com.wanted.momocity.enrollment.domain.exception.DuplicateEnrollmentExcept
 import com.wanted.momocity.enrollment.domain.exception.InvalidEnrollmentLectureStatusException;
 import com.wanted.momocity.enrollment.domain.model.Enrollment;
 import com.wanted.momocity.enrollment.domain.repository.EnrollmentRepository;
-import com.wanted.momocity.enrollment.infrastructure.metrics.BuildingMetrics;
 import com.wanted.momocity.enrollment.infrastructure.metrics.EnrollmentMetrics;
 import com.wanted.momocity.lecture.domain.model.LectureStatus;
 import lombok.RequiredArgsConstructor;
@@ -26,7 +25,6 @@ import org.springframework.transaction.annotation.Transactional;
 public class EnrollmentCommandService implements EnrollmentCommandUseCase {
 
     private final EnrollmentMetrics enrollmentMetrics;
-    private final BuildingMetrics buildingMetrics;
 
     // 수강신청 저장소
     private final EnrollmentRepository enrollmentRepository;
@@ -42,56 +40,60 @@ public class EnrollmentCommandService implements EnrollmentCommandUseCase {
 
     @Override
     public Enrollment createEnrollment(CreateEnrollmentCommand command) {
-        // Authorization 토큰에서 꺼낸 email로 학생 ID를 조회
-        Long userId = studentAccountPort.getStudentId(command.studentId());
 
-        // 수강신청 대상 강의의 현재 상태를 조회
-        LectureStatus lectureStatus = enrollmentLecturePort.getLectureStatus(command.lectureId());
+        log.info("수강신청 시작 - lectureId={}", command.lectureId());
 
-        // 같은 학생이 같은 강의를 이미 수강신청했는지 확인
-        boolean alreadyEnrolled = enrollmentRepository.existsByUserIdAndLectureId(
-                userId,
-                command.lectureId()
-        );
+            // Authorization 토큰에서 꺼낸 email로 학생 ID를 조회
+            Long userId = studentAccountPort.getStudentId(command.studentId());
 
-        // 이미 수강신청한 강의라면 중복 수강신청 예외를 발생
-        if (alreadyEnrolled) {
-            throw new DuplicateEnrollmentException("이미 수강신청한 강의입니다.");
-        }
+            // 수강신청 대상 강의의 현재 상태를 조회
+            LectureStatus lectureStatus = enrollmentLecturePort.getLectureStatus(command.lectureId());
 
-        // ACTIVE 상태의 강의만 수강신청할 수 있음
-        if (lectureStatus != LectureStatus.ACTIVE) {
-            throw new InvalidEnrollmentLectureStatusException("진행 중인 강의만 수강신청할 수 있습니다.");
-        }
+            // 같은 학생이 같은 강의를 이미 수강신청했는지 확인
+            boolean alreadyEnrolled = enrollmentRepository.existsByUserIdAndLectureId(
+                    userId,
+                    command.lectureId()
+            );
 
-        // 수강신청 도메인 객체를 생성
-        Enrollment enrollment = Enrollment.create(
-                userId,
-                command.lectureId()
-        );
+            // 이미 수강신청한 강의라면 중복 수강신청 예외를 발생
+            if (alreadyEnrolled) {
+                throw new DuplicateEnrollmentException("이미 수강신청한 강의입니다.");
+            }
 
-        // 수강신청 정보를 저장
-        Enrollment savedEnrollment = enrollmentRepository.save(enrollment);
+            // ACTIVE 상태의 강의만 수강신청할 수 있음
+            if (lectureStatus != LectureStatus.ACTIVE) {
+                throw new InvalidEnrollmentLectureStatusException("진행 중인 강의만 수강신청할 수 있습니다.");
+            }
 
-        // 강의별 수강 신청 누적 횟수 기록 -> actuator(인기강의용)
-        enrollmentMetrics.recordEnrollmentCreated(command.lectureId());
-        // 건물 획득 누적
-        buildingMetrics.recordBuildingAcquired();
+            // 수강신청 도메인 객체를 생성
+            Enrollment enrollment = Enrollment.create(
+                    userId,
+                    command.lectureId()
+            );
 
-        // 수강신청 완료 후 강사 자동 친구 추가를 위해 이벤트를 발행
-        eventPublisher.publishEvent(new EnrollmentCompletedEvent(
-                savedEnrollment.getUserId(),
-                savedEnrollment.getLectureId()
-        ));
+            // 수강신청 정보를 저장
+            Enrollment savedEnrollment = enrollmentRepository.save(enrollment);
 
-        // 수강신청 완료와 이벤트 발행 여부를 추적하기 위한 로그
-        // 수강신청 후 강사 자동 친구 추가 이벤트가 이어지므로 enrollmentId, userId, lectureId를 남긴다.
-        log.info("수강신청 완료 - enrollmentId={}, userId={}, lectureId={}",
-                savedEnrollment.getId(),
-                savedEnrollment.getUserId(),
-                savedEnrollment.getLectureId());
+            // 강의별 수강 신청 누적 횟수 기록 -> actuator(인기강의용)
+            enrollmentMetrics.recordEnrollmentCreated(command.lectureId());
+            // 건물 획득(수강 신청) 누적
+            enrollmentMetrics.recordBuildingAcquired();
 
-        // 저장된 수강신청 도메인 객체를 반환
-        return savedEnrollment;
+            // 수강신청 완료 후 강사 자동 친구 추가를 위해 이벤트를 발행
+            eventPublisher.publishEvent(new EnrollmentCompletedEvent(
+                    savedEnrollment.getUserId(),
+                    savedEnrollment.getLectureId()
+            ));
+
+            // 수강신청 완료와 이벤트 발행 여부를 추적하기 위한 로그
+            // 수강신청 후 강사 자동 친구 추가 이벤트가 이어지므로 enrollmentId, userId, lectureId를 남긴다.
+            log.info("수강신청 완료 - enrollmentId={}, userId={}, lectureId={}",
+                    savedEnrollment.getId(),
+                    savedEnrollment.getUserId(),
+                    savedEnrollment.getLectureId());
+
+            // 저장된 수강신청 도메인 객체를 반환
+            return savedEnrollment;
+
     }
 }
